Code Refactoring - Inner Classes in different files
